[🔨refactor] 메뉴 삭제 기능 관련 함수 분리
